{{cookiecutter.service_name}}/requirements-test.txt
pytest==7.1.2
pytest-asyncio==0.18.3
requests==2.28.0
pytest-async-sqlalchemy==0.2.0
